==> analysis/experiments_2-3/analyses/3_dichotic.Rmd <==
---
title: "Experiment 3: Dichotic Listening Analyses"
pagetitle: "exp 3 | dichotic listening analyses"
author: "Owen Morgan"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: html_document
---

<style type="text/css">
  body{
  font-family: Avenir;
  font-size: 12pt;
}
</style>

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.align = "center", fig.width = 9,
                      fig.height = 6, results = "asis")
options(knitr.kable.NA = "")

cli.progress_show_after <- 0

## Do not use scientific notation until 9 decimal places.
options(scipen = 9, digits = 9)

# Suppress summarise info
options(dplyr.summarise.inform = FALSE)
```

```{r lib}
library(here)
library(ppcor) ## Masks select()
library(tidyverse)
select <- dplyr::select
library(cli) # For printing error messages

library(lme4)
library(emmeans)
library(broom)
library(gt)

library(ggpubr) ## functions for plotting SD, SEM, CI.
library(ggbeeswarm)
library(patchwork)
library(ggh4x) ## for nested facets
library(glue)

library(knitr) # For include_graphics

source(here::here("lib", "util.R")) # interaction_stats, ...
source(here::here("lib", "demographics.R"))
source(here::here("lib", "plots-jepg.R"))

EXPERIMENT_DIR <- "experiments_2-3"
source(here::here(EXPERIMENT_DIR, "lib", "load_process", "load_process.R"))

## Colors for plots
plot_blue <- c("#337aB7")
plot_colors <- c("#4477AA",
                 "#EE6677",
                 "#CCBB44",
                 "#66CCEE",
                 "#AA3377",
                 "#228833",
                 "#BBBBBB")
```

```{r config}
proc_dir <- set_or_make_dir(label = "proc",
  here::here(EXPERIMENT_DIR, "data", "proc_exp_n1450")
)
fig_dir <- set_or_make_dir(label = "fig",
  here::here(EXPERIMENT_DIR, "figures")
)
manual_cache_dir <- set_or_make_dir(label = "manual cache",
 here::here(EXPERIMENT_DIR, "manual_cache")
)

experiment_label <- "Experiment 3"

## Emmeans settings (use Satterthwaire approx. instead of default t-to-z for large df)
## See: Luke (2017). Evaluating significance in linear mixed-effects models in R
DF_METHOD <- "satterthwaite"
LMERTEST_LIMIT <- 500000

## Figure settings
DEFAULT_HEIGHT = 3
DEFAULT_WIDTH = 5

use_cached_models <- F
use_cached_figs <- F
use_cached_proc_data <- F
```

# {.tabset}
```{r load_data}
## Load "the data" with all subjects & AAH trials.
aah_long <- load_aah_long(proc_dir)

## Load summary data table
## with all subjects.
aah_summary_all <- read_tsv(here(proc_dir, "aah_summary.tsv"))

## Subset to included subjects
aah_summary <- aah_summary_all |>
    filter(exclude_any == 0) |>
    select(-starts_with("exclude"))
```

```{r calculate_variables}
#### Prepare "the data" (aah_long) for all analyses:
## Filter out practice trials, absent trials, and excluded subjects
## This data (all present trials, correct or incorrect) will be used for
## accuracy analyses
cli::cli_alert_info("Filtering summary data for analysis (all exclusions).")
cli::cli_bullets(c(" " = "aah_summary_all -> aah_summary"))
aah <- aah_long |>
    filter(exclude_any == 0) |> # (Only include ppts with both valid DL and Navon data)
    filter(block_type == "main" & target_present == "yes") |>
    select(-starts_with("exclude")) |>
    ## Recode "level", "target" for nicer printing.
    mutate(level = recode(level, global = "Global", local = "Local")) |>
    mutate(target = recode(target, square = "Square", rectangle = "Rectangle")) |>
  mutate(
    handedness_extremes = case_when(ehi == -100 ~ "Left",
                                    ehi == 100 ~ "Right",
                                    ehi > -100 &
                                      ehi < 100 ~ "Mixed")
  ) |> rename(shape = target,
              clickhand = DL_i1_clickhand)

## For RT analyses, prepare dataset with only correct, present trials.
## In our RT model, we only care about correct responses to present trials.
aah_correct <- aah |> filter(correct == T)

## Relevel field and level with RVF first (unintuitive for plotting),
## so that emmeans will show a positive number for LVF global bias.
aah_for_rt_model <- aah_correct |>
  mutate(level = level |> factor(levels = c("Global", "Local")),
         field = field |> factor(levels = c("RVF", "LVF")))

##Count participants, with REA, LEA, or none
dic_data <- aah_summary |> 
  mutate(ear_advantage = case_when(
    (DL_lat <= 10 & DL_lat >= -10) ~ "none",
    DL_lat < 10 ~ "LEA",
    DL_lat > 10 ~ "REA"
  )) |> 
  ## Recode clickhand
  mutate(clickhand = (DL_i1_clickhand)) |> 
  ## Calculate RVF>LVF bias for Global, local
  ## A positive number means RVF bias
  rename(rt_Global_LVF = rt_global_LVF,
         rt_Local_LVF = rt_local_LVF,
         rt_Global_RVF = rt_global_RVF,
         rt_Local_RVF = rt_local_RVF,
         acc_Global_LVF = acc_global_LVF,
         acc_Local_LVF = acc_local_LVF,
         acc_Global_RVF = acc_global_RVF,
         acc_Local_RVF = acc_local_RVF,
         ) |> 
  ##
  #### RT
  ##
  mutate(rt_Global_hemifield_bias = rt_Global_LVF - rt_Global_RVF) |> 
  mutate(rt_Local_hemifield_bias = rt_Local_LVF - rt_Local_RVF) |> 
  ## Calculate frequency asymmetry
  ## A positive number means relative RVF local bias
  mutate(rt_freq_asym = rt_Local_hemifield_bias - rt_Global_hemifield_bias) |> 
  ##
  #### Accuracy
  ##
  mutate(acc_Global_hemifield_bias = acc_Global_RVF - acc_Global_LVF) |> 
  mutate(acc_Local_hemifield_bias = acc_Local_RVF - acc_Local_LVF) |> 
  ## Calculate frequency asymmetry
  ## A positive number means relative RVF local bias
  mutate(acc_freq_asym = acc_Local_hemifield_bias - acc_Global_hemifield_bias)

cor_data <- dic_data


## RT data
if (use_cached_proc_data == FALSE) {
rt_subject <- aah_correct |> group_by(subject, field, level, handedness) |>
  summarize(rt = median(rt))
  saveRDS(rt_subject, here(manual_cache_dir, "rt_subject.rds"))
} else {
  rt_subject <- readRDS(here(manual_cache_dir, "rt_subject.rds"))
}


## Prepare subject-level LVF Global bias summary (RT)
if (use_cached_proc_data == FALSE) {
rt_1 <- rt_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = rt) |>
  mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
  mutate(all_one_group = "all_one_group")
  saveRDS(rt_1, here(manual_cache_dir, "rt_1.rds"))
} else {
  rt_1 <- readRDS(here(manual_cache_dir, "rt_1.rds"))
}
```
## Demographics {.tabset .tabset-pills}

Demographics for all included participants.
<br>

```{r demo_summary}
rt_demo <- demo_summary_table(aah_summary)
rt_demo |> pretty_table() |> tab_header(title = "Demographics", subtitle = "Summary")
```
<br>

```{r}
hand_demo_summary_table(aah_summary) |> pretty_table() |>
  tab_footnote("Left: (EHI <= -40)  |  Mixed: (-40 < EHI < 40)  |  Right: (EHI >= 40)")
```
<br>
```{r ehi_distribution_plot_purple1}
middle_purple1 <- "#996F91" ## In between plot colors one and two
fig_path_var <- here(fig_dir, "demo_ehi_dots_purple.png")

if (use_cached_figs == F) {
  ehi_plot_data <- aah_summary |> group_by(ehi) |> summarize(n = n())
  g <- gg_ehi_dotarea(ehi_plot_data, plot_color = middle_purple1,
                      cutoff = 40, cutoff_label_ypos = 105)
  ggsave(fig_path_var, g, "png", height = 3, width = 6)
}

include_graphics(fig_path_var)
```


## DL ~ EHI (cont.) (primary) {.tabset .tabset-pills}

Does EHI predict language laterality? (Analyses comparable to Packheiser et al., 2020)

```{r}
## Relevel so that emmeans shows a positive number for RVF bias
aah_cont <- aah_for_rt_model |> 
  mutate(level = level |> factor(levels = c("Local", "Global")),
         field = field |> factor(levels = c("LVF", "RVF")))
```

### Stats

#### Model: DL (continuous) ~ EHI

Does handedness correlate with DL?

```{r}
## Use Cor.test
cor_test <- cor.test(cor_data$ehi, cor_data$DL_lat)
cor_test |> tidy() |> format_p.value() |> 
  pretty_table("Correlation (rho) between EHI and language laterality (DL ~ ehi)")
```
<br>
```{r}
## Use a linear model
lm_dl_hand_cont <- lm(data = cor_data, formula = DL_lat ~ ehi)
## Use emtrends to estimate change in DL units per EHI unit, with confidence intervals.
rt_emt <- emtrends(lm_dl_hand_cont, ~ ehi, var = "ehi",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)

rt_emt |> summary(infer = T) |> tidy() |> format_p.value() |> 
  pretty_table("Model: DL ~ ehi") |> 
  tab_footnote(footnote = "Change in DL units (-100 to 100) per EHI unit (-100 to 100). Multiply by 200 to estimate the difference between strong left and right handers",
               locations = cells_column_labels(columns = ehi.trend))
```
<br>
When limiting analysis to participants who clicked with their right hand?
<br>
```{r}
## Use Cor.test
model_data <- cor_data |> filter(clickhand == 100)
cor_test <- cor.test(model_data$ehi, model_data$DL_lat)
cor_test |> tidy() |> format_p.value() |> 
  pretty_table("Correlation (rho) between EHI and language laterality (DL ~ ehi; right clickers)")
```
<br>
```{r}
## Use a linear model
lm_dl_hand_cont <- lm(data = model_data, formula = DL_lat ~ ehi)
## Use emtrends to estimate change in DL units per EHI unit, with confidence intervals.
rt_emt <- emtrends(lm_dl_hand_cont, ~ ehi, var = "ehi",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)

rt_emt |> summary(infer = T) |> tidy() |> format_p.value() |> 
  pretty_table("Model: DL ~ ehi (right clickers)") |> 
  tab_footnote(footnote = "Change in DL units (-100 to 100) per EHI unit (-100 to 100). Multiply by 200 to estimate the difference between strong left and right handers",
               locations = cells_column_labels(columns = ehi.trend))
```


#### Model: DL (continuous) ~ clickhand

Does clickhand correlate with DL?

```{r}
## Use Cor.test
cor_test <- cor.test(cor_data$clickhand, cor_data$DL_lat)
cor_test |> tidy() |> format_p.value() |> 
  pretty_table("Correlation (rho) between clickhands and language laterality (DL ~ clickhand)")
```
<br>
```{r}
## Use a linear model
lm_dl_hand_cont <- lm(data = cor_data, formula = DL_lat ~ clickhand)
## Use emtrends to estimate change in DL units per EHI unit, with confidence intervals.
rt_emt <- emtrends(lm_dl_hand_cont, ~ clickhand, var = "clickhand",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)
rt_emt |> summary(infer = T) |> tidy() |> format_p.value() |> 
  pretty_table("Model: DL ~ ehi") |> 
  tab_footnote(footnote = "Change in DL units (-100 to 100) per clickhand unit (-100 to 100). Multiply by 200 to estimate the difference between strong left and right clickers",
               locations = cells_column_labels(columns = clickhand.trend))
```
<br>

#### Model: clickhand ~ ehi

Does handedness predict clickhand?

```{r}
## Use Cor.test
cor_test <- cor.test(cor_data$clickhand, cor_data$ehi)
cor_test |> tidy() |> format_p.value() |> 
  pretty_table("Correlation (rho) between clickhand and handedness (clickhand ~ ehi)")
```
<br>
```{r}
## Use a linear model
lm_dl_hand_cont <- lm(data = cor_data, formula = clickhand ~ ehi)
## Use emtrends to estimate change in DL units per EHI unit, with confidence intervals.
rt_emt <- emtrends(lm_dl_hand_cont, ~ ehi, var = "ehi",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)
rt_emt |> summary(infer = T) |> tidy() |> format_p.value() |> 
  pretty_table("Model: DL ~ ehi") |> 
  tab_footnote(footnote = "Change in clickhand units (-100 to 100) per EHI unit (-100 to 100). Multiply by 200 to estimate the difference between strong left and right handers",
               locations = cells_column_labels(columns = ehi.trend))
```
<br>

#### Model: DL (continuous) ~ EHI + clickhand

Does EHI predict DL, even when controlling for clickhand?
<br>
```{r}
MODEL_DESCRIPTION <- "Model: DL (cont.) ~ EHI + clickhand"
## Find rho for partial correlation

library(ppcor)
## Adapted from: https://stackoverflow.com/questions/62652711/confidence-interval-for-partial-correlations
pcor_ci.test <-
function (x, y, z, method = c("pearson", "kendall", "spearman"), conf.level = 0.95, ...) {
    d1 <- deparse(substitute(x))
    d2 <- deparse(substitute(y))
    d3 <- deparse(substitute(z))
    data.name <- paste0(d1, " and ", d2, "; controlling: ", d3)
    method <- match.arg(method)
    Method <- paste0("Partial correlation (", method, ")")
    alternative <- "true partial correlation is not equal to 0"

    x <- as.vector(x)
    y <- as.vector(y)
    z <- as.data.frame(z)
    xyz <- data.frame(x, y, z)
    pcor <- ppcor::pcor(xyz, method = method)
    estimate <- pcor$est[1, 2]
    p.value <- pcor$p.value[1, 2]
    parameter <- c(n = pcor$n, gp = pcor$gp)
    statistic <- c(Stat = pcor$statistic[1, 2])

    fit1 <- lm(x ~ z, data = xyz)
    fit2 <- lm(y ~ z, data = xyz)
    cortest <- cor.test(resid(fit1), resid(fit2), method = method, conf.level = conf.level, ...)
    ci <- cortest$conf.int

    out_tbl <- 
      tibble(
         statistic = statistic,
        n = pcor$n,
        gp = pcor$gp,
        p.value = p.value,
        estimate = c(partial.cor = estimate),
        LCL = ci[[1]],
        UCL = ci[[2]],
        alternative = alternative,
        method = Method,
        data.name = data.name
      )
    
    return(out_tbl)
}

p_cor <- pcor_ci.test(cor_data$DL_lat, cor_data$ehi, cor_data$clickhand)
p_cor |> format_p.value() |> 
    pretty_table("Partial correlation (rho) between DL and ehi (DL ~ ehi * clickhand)") |> 
  tab_footnote(footnote = "t-statistic",
               locations = cells_column_labels(columns = statistic))
```
<br>
```{r}
## use a linear model (DL ~ ehi + clickhand)
## Effect of EHI on DL

lm_dl_hand_cont_click <- lm(data = cor_data, formula = DL_lat ~ ehi + clickhand)
rt_emt <- emtrends(lm_dl_hand_cont_click, ~ ehi, var = "ehi",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)

rt_emt |> summary(infer = T) |> tidy() |> format_p.value() |> 
  pretty_table("Model: DL ~ ehi + clickhand") |> 
  tab_footnote(footnote = "Change in DL units (-100 to 100) per EHI unit (-100 to 100). Multiply by 200 to estimate the difference between strong left and right handers",
               locations = cells_column_labels(columns = ehi.trend))
```
<br>

Is the model with clickhand better than the one without? (Could clickhand drive or inflate the effect of handedness on DL?)

```{r}
## Compare model with and without clickhand
## Could clickhand drive or inflate the effect of handedness on DL?
lm_no_click <- update(lm_dl_hand_cont_click, . ~ . - clickhand)
click_anova <- anova(lm_dl_hand_cont_click, lm_no_click)

click_anova |> 
  as_tibble() |> 
  rename(p.value = `Pr(>F)`) |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Does clickhand predict language laterality, over and above EHI?", 
             subtitle = "ANOVA: compare models with vs. without interaction term") |> 
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "F-test (two-sided? https://daniellakens.blogspot.com/2016/04/one-sided-f-tests-and-halving-p-values.html)",
               locations = cells_column_labels(columns = p.value))

```
<br>
```{r}
## use a linear model (DL ~ ehi + clickhand)
## Effect of clickhand on DL

lm_dl_hand_cont_click <- lm(data = cor_data, formula = DL_lat ~ ehi + clickhand)
rt_emt <- emtrends(lm_dl_hand_cont_click, ~ clickhand, var = "clickhand",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)

rt_emt |> summary(infer = T) |> tidy() |> format_p.value() |> 
  pretty_table("Model: DL ~ ehi + clickhand") |> 
  tab_footnote(footnote = "Change in DL units (-100 to 100) per clickhand unit (-100 to 100). Multiply by 200 to estimate the difference between strong left and right handers",
               locations = cells_column_labels(columns = clickhand.trend))
```
<br>

#### Model: DL (continuous) ~ EHI * clickhand
<br>
```{r}
## use a linear model (DL ~ ehi * clickhand)
lm_dl_hand_cont_click <- lm(data = cor_data, formula = DL_lat ~ ehi*clickhand)
lm_dl_hand_cont_click |> emmeans(~ ehi,
                          lmer.df = DF_METHOD,
                          lmerTest.limit = LMERTEST_LIMIT)
lm_dl_hand_cont_click |> tidy() |> format_p.value() |> 
  pretty_table("Model: DL ~ ehi * clickhand") |> 
  tab_footnote(footnote = "Change in DL units (-100 to 100) per EHI unit (-100 to 100). Multiply by 200 to estimate the difference between strong left and right handers",
               locations = cells_column_labels(columns = estimate))
```
<br>

### Plots

```{r}
middle_purple1 <- "#996F91" ## In between plot colors one and two
middle_purple2 <- "#886e92" ## In between plot colors one and two
plot_color <- middle_purple1

h_plot_colors <- c(plot_colors[[1]], middle_purple1, plot_colors[[2]])
```

```{r}
## Plot DL ~ EHI (by EHI bin)
fig_path_var <- here(fig_dir, "E3_cor.png")

## Find model estimates, n subjects
lm_dl_hand_cont <- lm(data = cor_data, formula = DL_lat ~ ehi)

n_subjects <- cor_data |> group_by(subject) |> count() |> group_by(subject) |> pull(subject) |> length() |> 
  format(nsmall = 0, big.mark = ",")

line_data <- lm_dl_hand_cont |> 
  ref_grid(at = list(ehi = seq(-100, 100, 12.5)),
           lmer.df = DF_METHOD,
                        lmerTest.limit = LMERTEST_LIMIT
           ) |> 
  emmeans(~ ehi) |> 
  summary(infer = T) |> 
  as_tibble()


## Goal: a tibble with one row per subject, columns: ehi, dv
subject_plot <- cor_data |> 
  dplyr::select(subject, ehi, DL_lat) |> 
  rename(dv = DL_lat)

  g <-
    gg_rt_1_line(
      title = "Language laterality by handedness (continuous)",
      y_title = "Dichotic listening laterality score",
      rt_subject_plot = subject_plot,
      plot_colors = h_plot_colors,
      plot_color = plot_color,
      group_by_level = F,
      direction_labels = list(
        up = "Right ear (left hemisphere) dominance",
        down = "Left ear (right hemisphere) dominance"),
      direction_labels_pos = list(
        up = 43,
        down = 0),
      ylims = list(
        upper = 45,
        lower = -10),
      ybreaks = list(
        major = 10,
        minor = 5)
    ) |>
    gg_style_cor_bin()
  g <- g +
  geom_line(data = line_data,
            aes(y = emmean)) +
    geom_ribbon(data = line_data,
                inherit.aes = F,
              aes(x = ehi, ymin = lower.CL, ymax = upper.CL),
              fill = plot_color,
              alpha = .2)
  
  
  g <- g + labs(subtitle = glue("Experiment 3 (n = ", n_subjects, ")")) +
    theme(plot.title = element_text(hjust = .5),
          plot.subtitle = element_text(hjust = .5, face = "italic")
    )
  
  ggsave(fig_path_var, g, "png", height = 4, width = 4)
  
include_graphics(fig_path_var)
```

<br>

```{r}
#### Limit analysis to participants who clicked with their right hand
## Plot DL ~ EHI (by EHI bin)
## Follow: continuous plot code
fig_path_var <- here(fig_dir, "E3_cor_rightclick.png")

## Find model estimates, n subjects
model_data <- cor_data |> filter(clickhand == 100)
lm_dl_hand_cont <- lm(data = model_data, formula = DL_lat ~ ehi)

n_subjects <- model_data |> group_by(subject) |> count() |> group_by(subject) |> pull(subject) |> length() |> 
  format(nsmall = 0, big.mark = ",")

line_data <- lm_dl_hand_cont |> 
  ref_grid(at = list(ehi = seq(-100, 100, 12.5)),
           lmer.df = DF_METHOD,
                        lmerTest.limit = LMERTEST_LIMIT
           ) |> 
  emmeans(~ ehi) |> 
  # contrast(interaction = c("consec"), by = c("ehi", "level"), level = .95) |> 
  summary(infer = T) |> 
  as_tibble()


## Goal: a tibble with one row per subject, columns: ehi, dv
subject_plot <- model_data |> 
  dplyr::select(subject, ehi, DL_lat) |> 
  rename(dv = DL_lat)

  g <-
    gg_rt_1_line(
      title = "Language laterality by handedness (continuous)",
      y_title = "Dichotic listening laterality score",
      rt_subject_plot = subject_plot,
      plot_colors = h_plot_colors,
      plot_color = plot_color,
      group_by_level = F,
      direction_labels = list(
        up = "Right ear (left hemisphere) dominance",
        down = "Left ear (right hemisphere) dominance"),
      direction_labels_pos = list(
        up = 43,
        down = 0),
      ylims = list(
        upper = 45,
        lower = -10),
      ybreaks = list(
        major = 10,
        minor = 5)
    ) |>
    gg_style_cor_bin()
  g <- g +
  geom_line(data = line_data,
            aes(y = emmean)) +
    geom_ribbon(data = line_data,
                inherit.aes = F,
              aes(x = ehi, ymin = lower.CL, ymax = upper.CL),
              fill = plot_color,
              alpha = .2)
  
  
  g <- g + labs(subtitle = glue("Participants who clicked with their right hand (n = ", n_subjects, ")")) +
    theme(plot.title = element_text(hjust = .5),
          plot.subtitle = element_text(hjust = .5, face = "italic")
    )
  
  ggsave(fig_path_var, g, "png", height = 4, width = 4)
  
include_graphics(fig_path_var)
```


## DL (cont.) ~ handedness (cat.) (primary) {.tabset .tabset-pills}

Does handedness (categorical) predict language laterality? (Analyses comparable to Karlsson et al. 2023).

### Stats

Model: DL (categorical) ~ handedness
```{r}
## Test whether DL_lat is significantly greater than zero, for each group
dic_data_sub <- dic_data |> filter(handedness %in% c("Left", "Right"))
                                   
lm_hand_dic <- lm(data = dic_data_sub, formula = DL_lat ~ handedness)
emmeans(lm_hand_dic, ~ handedness) |> 
  summary(infer = T) |> 
  as_tibble() |> 
  format_p.value() |> 
  pretty_table("Language laterality by handedness")
```
<br>
```{r}
## Test whether groups differ in their degree of language laterality
emmeans(lm_hand_dic, ~ handedness) |> 
  contrast("pairwise", infer = T) |> 
  as_tibble() |> 
  format_p.value() |> 
  pretty_table("Difference in language laterality by handedness group")
```
<br>
Does the effect persist within participants who always clicked with their right hand?
<br>
```{r}
## Test whether DL_lat is significantly greater than zero, for each group
dic_data_sub <- dic_data |> filter(handedness %in% c("Left", "Right")) |> filter(clickhand == 100)
                                   
lm_hand_dic <- lm(data = dic_data_sub, formula = DL_lat ~ handedness)
emmeans(lm_hand_dic, ~ handedness) |> 
  summary(infer = T) |> 
  as_tibble() |> 
  format_p.value() |> 
  pretty_table("Language laterality by handedness (right clickers)")
```
<br>
```{r}
## Test whether groups differ in their degree of language laterality
emmeans(lm_hand_dic, ~ handedness) |> 
  contrast("pairwise", infer = T) |> 
  as_tibble() |> 
  format_p.value() |> 
  pretty_table("Difference in language laterality by handedness group (right clickers)")
```
<br>
Model: DL (categorical) ~ handedness + clickhand
<br>
```{r}
## Test whether DL_lat is significantly greater than zero, for each group
dic_data_sub <- dic_data |> filter(handedness %in% c("Left", "Right"))
                                   
lm_hand_dic <- lm(data = dic_data_sub, formula = DL_lat ~ handedness + clickhand)
emmeans(lm_hand_dic, ~ handedness) |> 
  summary(infer = T) |> 
  as_tibble() |> 
  format_p.value() |> 
  pretty_table("Language laterality by handedness")
```
<br>
```{r}
## Test whether groups differ in their degree of language laterality
emmeans(lm_hand_dic, ~ handedness) |> 
  contrast("pairwise", infer = T) |> 
  as_tibble() |> 
  format_p.value() |> 
  pretty_table()
```
<br>

### Plots

```{r}
## Plot language laterality distributions (right and left)
fig_path <- here(fig_dir, "E3_density.png")

if (use_cached_figs == FALSE) {
  
  ## Extract n's (MANUALLY COUNT N IN DATAFRAME USED IN MODEL)
  n_left <- dic_data |> group_by(subject, handedness) |>  count() |>
    group_by(handedness) |> count() |>
    filter(handedness == "Left") |> pull(n)

  n_right <- dic_data |> group_by(subject, handedness) |>  count() |>
    group_by(handedness) |> count() |>
    filter(handedness == "Right") |> pull(n)
  
  n_subjects = list(right = n_right,
                      left = n_left)
  
  n_subjects_total <- n_left + n_right
  
  handedness_labeller <- c(
      Right = str_c("Right (n = ", n_subjects$right, ")"),
      Left = str_c("Left (n = ", n_subjects$left, ")")
    )
  
  g_density <- dic_data |> filter(handedness %in% c("Right", "Left")) |> 
    ggplot(aes(x = DL_lat, fill = handedness)) +
    geom_vline(xintercept = 0, color = "gray50", linewidth = .5) +
    geom_density(alpha = .3, show.legend = T) +
    labs(x = "Language Laterality", y = "",
         title = "Language laterality by handedness (EHI cut at +/-40)") +
        scale_fill_manual(values = plot_colors[c(1, 2)],
                          labels = c(handedness_labeller[["Left"]],
                                  handedness_labeller[["Right"]])
                          ) +
    scale_color_manual(values = plot_colors[c(1, 2)])
    # facet_wrap(~handedness, ncol = 1)
  g_density <- g_density |> gg_style() +
    theme(aspect.ratio = 1 / 2,
          axis.text.y = element_blank(),
          plot.title = element_text(hjust = 0.5)
          ) +
    theme(legend.position = c(.2, .8)) +
    guides(fill = guide_legend(title = "Handedness")) +
    labs(subtitle = glue("Experiment 3 (n = ", n_subjects_total, ")")) +
    theme(plot.title = element_text(hjust = .5)) +
    theme(plot.subtitle = element_text(hjust = .5, face = "italic"))
  
  ggsave(fig_path,
         g_density,
         "png",
         height = 3,
         width = 6)
}

include_graphics(fig_path)
```

```{r}
## check: clickhand == right
## (not enough participants who clicked with left to run that - only 5 right handers)
fig_path <- here(fig_dir, "E3_density_rightclick.png")

if (use_cached_figs == FALSE) {
  
  model_data <- dic_data |> filter(clickhand == 100)
  ## Extract n's (MANUALLY COUNT N IN DATAFRAME USED IN MODEL)
  n_left <- model_data |> group_by(subject, handedness) |>  count() |>
    group_by(handedness) |> count() |>
    filter(handedness == "Left") |> pull(n)

  n_right <- model_data |> group_by(subject, handedness) |>  count() |>
    group_by(handedness) |> count() |>
    filter(handedness == "Right") |> pull(n)
  
  n_subjects = list(right = n_right,
                      left = n_left)
  
  n_subjects_total <- n_left + n_right
  
  handedness_labeller <- c(
      Right = str_c("Right (n = ", n_subjects$right, ")"),
      Left = str_c("Left (n = ", n_subjects$left, ")")
    )
  
  g_density <- model_data |> filter(handedness %in% c("Right", "Left")) |> 
    ggplot(aes(x = DL_lat, fill = handedness)) +
    geom_vline(xintercept = 0, color = "gray50", linewidth = .5) +
    geom_density(alpha = .3, show.legend = T) +
    labs(x = "Language Laterality", y = "",
         title = "Language laterality by handedness (EHI cut at +/-40)") +
        scale_fill_manual(values = plot_colors[c(1, 2)],
                          labels = c(handedness_labeller[["Left"]],
                                  handedness_labeller[["Right"]])
                          ) +
    scale_color_manual(values = plot_colors[c(1, 2)])
    # facet_wrap(~handedness, ncol = 1)
  g_density <- g_density |> gg_style() +
    theme(aspect.ratio = 1 / 2,
          axis.text.y = element_blank(),
          plot.title = element_text(hjust = 0.5)
    ) +
    theme(legend.position = c(.2, .8)) +
    guides(fill = guide_legend(title = "Handedness")) +
    labs(subtitle = glue("Participants who clicked with their right hand (n = ", n_subjects_total, ")")) +
    theme(plot.title = element_text(hjust = .5)) +
    theme(plot.subtitle = element_text(hjust = .5, face = "italic"))
  
  ggsave(fig_path,
         g_density,
         "png",
         height = 3,
         width = 6)
}

include_graphics(fig_path)
```


## Percentage REA/LEA ~ handedness


```{r}
##Count participants, with REA, LEA, or none
group_ns <- dic_data |> group_by(handedness) |> 
  summarize(n_group = n())
dic_data |> 
  full_join(group_ns, by = "handedness") |> 
  mutate(ear_advantage = case_when(
    (DL_lat <= 10 & DL_lat >= -10) ~ "none",
    DL_lat < 10 ~ "LEA",
    DL_lat > 10 ~ "REA"
  )) |> 
  group_by(handedness, ear_advantage) |> 
  summarize(
    n = n(),
    # group_n = first(n_group),
    pct = (n/first(n_group))*100
  ) |> 
  pretty_table("Language laterality distribution (cut at +/-10)")
```





## Continuous handedness (exploratory) {.tabset .tabset-pills}

### All language laterality {.tabset}

Do we find an interaction of field x level x handedness (continuous EHI score), when dichotic listening (continuous) is included in the model?
<br>
<br>

#### Stats
<br>
<br>
`rt_model_ehi <- lmer( rt ~ field*level*ehi*DL_lat*clickhand + (1 | subject) )`
<br>
<br>
```{r rt_model_ehil}
## Prepare data, so emmeans contrasts show
## Global bias as a positive number.
aah_for_rt_ehi_model <- aah_for_rt_model |>
  mutate(level = level |> factor(levels = c("Local", "Global")),
         field = field |> factor(levels = c("LVF", "RVF")))
```

```{r}
MODEL_DESCRIPTION <- "Model: field * level * ehi * DL_lat * clickhand + (1|subject) + (1|shape)"

if (use_cached_models == FALSE) {
  rt_mm_cont_click <- lmer(
    data = aah_cont,
    formula = rt ~
      field * level * ehi * DL_lat * clickhand
    + (1 | subject)
    + (1 | shape)
  )
  
  saveRDS(rt_mm_cont_click, here(manual_cache_dir, "rt_mm_cont_click.rds"))
} else if (use_cached_models == TRUE) {
  rt_mm_cont_click <- readRDS(here(manual_cache_dir, "rt_mm_cont_click.rds"))
}
```

```{r}
## FxLxH (emmeans)
## Estimate effect of field by level by EHI
rt_emt <- emtrends(rt_mm_cont_click, pairwise ~ field*level, var = "ehi",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                   )

## Report interaction estimate, with CI and p value.
rt_emt[[1]] |> contrast(interaction = c("consec"), var = "ehi") |> 
  summary(infer = T) |>
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Field by level by handedness (cont.) interaction (RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means local bias is stronger in RVF for right handers (as predicted by AAH), in ms per EHI unit (-100 to 100). Multiply this value by 200 to get the estimated difference in LVF global bias for strong left vs. right handers.",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```

<br>
```{r}
## FxH | L (emmeans)
## Report estimated effect of field, for each level, by ehi.
## Test whether interaction slope differs from zero, with emmeans

## This code estimates the effect of EHI on RT for each field*level:
rt_emt <- emtrends(rt_mm_cont_click, pairwise ~ field*level, var = "ehi",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)

## Report interaction estimate, with CI and p value.
rt_emt[[1]] |> contrast("pairwise", by = "level", var = "ehi") |> 
  summary(infer = T) |>
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = str_glue("{experiment_label}: Field by EHI (by Level) (RT)"), 
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means global bias is stronger in LVF for right handers (as predicted by AAH), in ms per EHI unit (-100 to 100). Multiply this value by 200 to get the estimated difference in LVF global bias for strong left vs. right handers.",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
br>

### Within REA {.tabset}

Do we find an interaction of field x level x handedness (continuous EHI score), when analysis is limited to participants with REA (DL > 10)?
<br>
<br>

#### Stats

<br>
<br>
`rt_model_ehi <- lmer( rt ~ field*level*ehi + (1 | subject) )`
<br>
<br>


Model: rt ~ field * level * ehi + (1|subject)
```{r}
MODEL_DESCRIPTION <- "Model: field * level * ehi + (1|subject)"

aah_cont_rea <- aah_cont |> left_join(dic_data) |>  filter(ear_advantage == "REA")

if (use_cached_models == FALSE) {
  rt_mm_cont_rea <- lmer(
    data = aah_cont_rea,
    formula = rt ~
      field * level * ehi 
    + (1 | subject)
  )
  
  saveRDS(rt_mm_cont_rea, here(manual_cache_dir, "rt_mm_cont_rea.rds"))
} else if (use_cached_models == TRUE) {
  rt_mm_cont_rea <- readRDS(here(manual_cache_dir, "rt_mm_cont_rea.rds"))
}
```
<br>
```{r}
## This code estimates the effect of EHI on RT for each field*level:
rt_emt <- emtrends(rt_mm_cont_rea, pairwise ~ field*level, var = "ehi",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)
## We want to compare the effect of EHI on RT for:
## (LVF Local - LVF Global) - (RVF local - RVF Global)
## Contrast indices: 2 - 5

## Report interaction estimate, with CI and p value.
rt_emt[[1]] |> contrast(interaction = c("consec"), var = "ehi") |> 
  summary(infer = T)|>
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "REA only: Field by level by handedness (cont.) interaction (RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means local bias is stronger in RVF for right handers (as predicted by AAH), in ms per EHI unit (-100 to 100). Multiply this value by 200 to get the estimated difference in LVF global bias for strong left vs. right handers.",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## Report EHI effect on bias by field, with CI and p-value
rt_emt$contrasts |>
  summary(infer = T) |> 
  as_tibble() |> 
  filter(contrast %in%
           c("LVF Local - RVF Local", "LVF Global - RVF Global")) |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "REA only: Field by level interaction by handedness (cont.; RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means more RVF bias for right handers, in ms per EHI unit (-100 to 100). Multiply by 200 to get the difference in ms between strong left and right handers",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))

```

## Categorical handedness (exploratory, EHI +/-40) {.tabset .tabset-pills}

```{r}
aah_cat <- aah_for_rt_model |> 
  full_join(dic_data) |> 
  filter(handedness %in% c("Right", "Left")) |> 
  filter(ear_advantage %in% c("REA", "LEA"))
```

### All language laterality {.tabset}

#### Stats
Reaction time is modeled as a linear effect of field, level, and handedness, using data from every target-present trial with a "go" response:
<br>
<br>
`lmer( rt ~ field*level*handedness*D_lat*clickhand + (1 | subject) )`
<br>
<br>
```{r}
## Count groups:
aah_cat |> group_by(handedness, subject) |> count() |> group_by(handedness) |> count() |> pretty_table()
```

```{r}
MODEL_DESCRIPTION <- "Model (yes clickhand): field * level * handedness * ear_advantage * clickhand + (1|subject) + (1|shape)"

if (use_cached_models == FALSE) {
  rt_mm_cat_click <- lmer(
    data = aah_cat,
    formula = rt ~
      field * level * handedness * ear_advantage * clickhand
    + (1 | subject)
    + (1 | shape)
  )
  
  saveRDS(rt_mm_cat_click, here(manual_cache_dir, "rt_mm_cat_click.rds"))
} else if (use_cached_models == TRUE) {
  rt_mm_cat_click <- readRDS(here(manual_cache_dir, "rt_mm_cat_click.rds"))
}
```
<br>
```{r}
## FxLxH
## Estimate effect of handedness on field x level with emmeans
## Use emmeans() to test 3-way interaction.
rt_emm <- emmeans(rt_mm_cat_click, ~ field * level * handedness,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  )
rt_interaction_emm <- rt_emm |> 
  contrast(interaction = c("consec")) |>
  summary(infer = T)

rt_interaction_emm |>
  as_tibble() |>
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "Field by level by handedness interaction (RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means RVF local bias is stronger in right handers (as predicted by AAH)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## FxL | H
## Estimate the effect of field by level for each handedness group
rt_emm <- emmeans(rt_mm_cat_click, ~field*level | handedness,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast(interaction = "consec") |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "LVF global bias by handedness bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## FxH | L
## Does handedness predict hemifield bias, at the local level?
## Do left handers show significantly difference local hemifield bias than right handers?
## Estimate the effect of field by level for each handedness group
rt_emm <- emmeans(rt_mm_cat_click, ~field*handedness | level,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast(interaction = c("consec"), by = c("level")) |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  arrange(level) |> 
  pretty_table() |>
  tab_header(title = "Effect of handedness on hemifield bias, by level (RT)") |>
  tab_footnote(footnote = "A positive number means more RVF bias for right handers",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## F | L | H
## Estimate the effect of field by level for each handedness group
rt_emm <- emmeans(rt_mm_cat_click, ~field | handedness + level,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast("revpairwise", by = c("handedness","level")) |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  arrange(handedness, level) |> 
  pretty_table() |>
  tab_header(title = "Hemifield bias by level by handedness bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>

#### Plots

```{r}
## Start over, with numeric y axis
## Add a line with CI from the model
fig_path_var <- here(fig_dir, "freq_hand_2bins_big_model.png")

rt_model <- rt_mm_cat_click

rt_emm <- rt_model |> emmeans( ~ field * level * handedness,
                               lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)

if (use_cached_figs == F) {
  ## Get estimates for right and left handers.
  emm_data_2 <- rt_emm |>
    contrast("pairwise", by = c("handedness", "level")) |>
    summary(infer = T, level = .95) |>
    as_tibble() |>
    ## Fix units
    mutate(
      estimate = estimate * -1,
      lower.CL = lower.CL * -1,
      upper.CL = upper.CL * -1
    )
  
  
  plot_data <- emm_data_2 |> mutate(dv = estimate)
  
  ## Extract n's (MANUALLY COUNT N IN DATAFRAME USED IN MODEL)
  n_left <-
    aah_cat |> group_by(handedness, subject) |> count() |> group_by(handedness) |> count() |> filter(handedness == "Left") |> pull(n)
  n_right <-
    aah_cat |> group_by(handedness, subject) |> count() |> group_by(handedness) |> count() |> filter(handedness == "Right") |> pull(n)
  
  g <- gg_rt_2_lmer(
    title = "Hemifield bias by level (EHI cut at +/-40)",
    plot_data,
    handedness_labeller = NULL,
    plot_colors = plot_colors,
   direction_labels = list(right = "RVF bias",
                          left = "LVF bias"),
  direction_labels_pos = list(right = 48,
                              left = -48),
    xlims = list(upper = 50,
                 lower = -50),
    xbreaks = list(major = 10,
                   minor = 5),
    n_subjects = list(right = n_right,
                      left = n_left)
  )
  
  ggsave(fig_path_var, g, "png", height = 3, width = 4)
}

include_graphics(fig_path_var)
```

## Within REA {.tabset}

#### Stats

Reaction time is modeled as a linear effect of field, level, and handedness, using data from every target-present trial with a "go" response:
<br>
<br>
`lmer( rt ~ field*level*handedness + (1 | subject) )`
<br>
<br>
```{r}
MODEL_DESCRIPTION <- "Model: field * level * handedness + (1|subject)"

aah_cat_rea <- aah_cat |> filter(ear_advantage == "REA")

## Count participants
aah_cat_rea |> group_by(handedness, subject) |> count() |> group_by(handedness) |> count() |> pretty_table()

if (use_cached_models == FALSE) {
  rt_mm_cat_rea <- lmer(
    data = aah_cat_rea,
    formula = rt ~
      field * level * handedness 
    + (1 | subject)
  )
  
  saveRDS(rt_mm_cat_rea, here(manual_cache_dir, "rt_mm_cat_rea.rds"))
} else if (use_cached_models == TRUE) {
  rt_mm_cat_rea <- readRDS(here(manual_cache_dir, "rt_mm_cat_rea.rds"))
}
```
<br>
```{r}
## FxLxH
## Estimate effect of handedness on field x level with emmeans
## Use emmeans() to test 3-way interaction.
rt_emm <- emmeans(rt_mm_cat_rea, ~ field * level * handedness,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  )
rt_interaction_emm <- rt_emm |> 
  contrast(interaction = c("consec")) |>
  summary(infer = T)

rt_interaction_emm |>
  as_tibble() |>
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "REA only: Field by level by handedness interaction (RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means RVF local bias is stronger in right handers (as predicted by AAH)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## FxL | H
## Estimate the effect of field by level for each handedness group
rt_emm <- emmeans(rt_mm_cat_rea, ~field*level | handedness,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT) |>
  contrast(interaction = "consec") |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "REA only: LVF Global bias by handedness bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>

```{r}
## FxH | L
## Does handedness predict hemifield bias, at the local level?
## Do left handers show significantly difference local hemifield bias than right handers?
## Estimate the effect of field by level for each handedness group
rt_emm <- emmeans(rt_mm_cat_rea, ~field*handedness | level,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast(interaction = c("consec"), by = c("level")) |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  arrange(level) |> 
  pretty_table() |>
  tab_header(title = "REA only: Effect of handedness on hemifield bias, by level (RT)") |>
  tab_footnote(footnote = "A positive number means more RVF bias for right handers",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>

```{r}
## F | H | L
## Estimate the effect of field by level for each handedness group
rt_emm <- emmeans(rt_mm_cat_rea, ~field | handedness + level,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast("revpairwise", by = c("handedness","level")) |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  arrange(handedness, level) |> 
  pretty_table() |>
  tab_header(title = "REA only: Hemifield bias by level by handedness bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>

## Categorical handedness (exploratory, EHI +/-100) {.tabset .tabset-pills}

### All language laterality {.tabset}

#### Stats

Reaction time is modeled as a linear effect of field, level, and handedness, using data from every target-present trial with a "go" response:
<br>
<br>
`lmer( rt ~ field*level*handedness*D_lat*clickhand + (1 | subject) )`
<br>
<br>
```{r}
## Count groups:
aah_cat_zoom <- aah_cat |> filter(handedness_extremes %in% c("Right", "Left"))
                                  
aah_cat_zoom |> group_by(handedness, subject) |> count() |> group_by(handedness) |> count() |> pretty_table()
```

```{r}
MODEL_DESCRIPTION <- "Model (yes clickhand): field * level * handedness_extremes * ear_advantage * clickhand + (1|subject) + (1|shape)"

if (use_cached_models == FALSE) {
  rt_mm_cat_click_zoom <- lmer(
    data = aah_cat_zoom,
    formula = rt ~
      field * level * handedness * ear_advantage * clickhand
    + (1 | subject)
    + (1 | shape)
  )
  
  saveRDS(rt_mm_cat_click_zoom, here(manual_cache_dir, "rt_mm_cat_click_zoom.rds"))
} else if (use_cached_models == TRUE) {
  rt_mm_cat_click_zoom <- readRDS(here(manual_cache_dir, "rt_mm_cat_click_zoom.rds"))
}
```
<br>
```{r}
## FxLxH
## Estimate effect of handedness on field x level with emmeans
## Use emmeans() to test 3-way interaction.
rt_emm <- emmeans(rt_mm_cat_click_zoom, ~ field * level * handedness,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  )
rt_interaction_emm <- rt_emm |> 
  contrast(interaction = c("consec")) |>
  summary(infer = T)

rt_interaction_emm |>
  as_tibble() |>
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "Field by level by handedness interaction (RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means RVF local bias is stronger in right handers (as predicted by AAH)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## FxL | H
## Estimate the effect of field by level for each handedness group
rt_emm <- emmeans(rt_mm_cat_click_zoom, ~field*level | handedness,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast(interaction = "consec") |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "LVF global bias by handedness bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## FxH | L
## Does handedness predict hemifield bias, at the local level?
## Do left handers show significantly difference local hemifield bias than right handers?
## Estimate the effect of field by level for each handedness group
rt_emm <- emmeans(rt_mm_cat_click_zoom, ~field*handedness | level,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast(interaction = c("consec"), by = c("level")) |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  arrange(level) |> 
  pretty_table() |>
  tab_header(title = "Effect of handedness on hemifield bias, by level (RT)") |>
  tab_footnote(footnote = "A positive number means more RVF bias for right handers",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## F | L | H
## Estimate the effect of field by level for each handedness group
rt_emm <- emmeans(rt_mm_cat_click_zoom, ~field | handedness + level,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast("revpairwise", by = c("handedness","level")) |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  arrange(handedness, level) |> 
  pretty_table() |>
  tab_header(title = "Hemifield bias by level by handedness bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>

## Categorical language (exploratory, EHI +/-40)

```{r}
aah_cat <- aah_for_rt_model |> 
  full_join(dic_data) |> 
  filter(handedness %in% c("Right","Left")) |> 
  filter(ear_advantage %in% c("REA", "LEA"))
```


#### Stats
Reaction time is modeled as a linear effect of field, level, and handedness, using data from every target-present trial with a "go" response:
<br>
<br>
`lmer( rt ~ field*level*handedness*ear_advantage*clickhand + (1 | subject) )`
<br>
<br>
```{r}
## Count groups:
aah_cat |> group_by(handedness, subject) |> count() |> group_by(handedness) |> count() |> pretty_table()
```

```{r}
MODEL_DESCRIPTION <- "Model (yes clickhand): field * level * handedness * ear_advantage * clickhand + (1|subject) + (1|shape)"

if (use_cached_models == FALSE) {
  rt_mm_cat_click <- lmer(
    data = aah_cat,
    formula = rt ~
      field * level * handedness * ear_advantage * clickhand
    + (1 | subject)
    + (1 | shape)
  )
  
  saveRDS(rt_mm_cat_click, here(manual_cache_dir, "rt_mm_cat_click.rds"))
} else if (use_cached_models == TRUE) {
  rt_mm_cat_click <- readRDS(here(manual_cache_dir, "rt_mm_cat_click.rds"))
}
```
<br>
```{r}
## FxLxH
## Estimate effect of language on field x level with emmeans
## Use emmeans() to test 3-way interaction.
rt_emm <- emmeans(rt_mm_cat_click, ~ field * level * ear_advantage,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  )
rt_interaction_emm <- rt_emm |> 
  contrast(interaction = c("consec")) |>
  summary(infer = T)

rt_interaction_emm |>
  as_tibble() |>
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "Field by level by language laterality interaction (RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means RVF local bias is stronger in right eared participants (as predicted by LAH)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## FxL | H
## Estimate the effect of field by level for each language group
rt_emm <- emmeans(rt_mm_cat_click, ~field*level | ear_advantage,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast(interaction = "consec") |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = "LVF global bias by language laterality bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## FxH | L
## Does language predict hemifield bias, at the local level?
## Do left earers show significantly difference local hemifield bias than right earers?
## Estimate the effect of field by level for each handedness group
rt_emm <- emmeans(rt_mm_cat_click, ~field*ear_advantage | level,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast(interaction = c("consec"), by = c("level")) |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  arrange(level) |> 
  pretty_table() |>
  tab_header(title = "Effect of language laterality on hemifield bias, by level (RT)") |>
  tab_footnote(footnote = "A positive number means more RVF bias for right handers",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## F | L | H
## Estimate the effect of field by level for each language group
rt_emm <- emmeans(rt_mm_cat_click, ~field | ear_advantage + level,
                  lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                  ) |>
  contrast("revpairwise", by = c("ear_advantage","level")) |>
  summary(infer = T, adj = "none")

rt_emm |>
  as_tibble() |> 
  format_p.value() |> 
  arrange(ear_advantage, level) |> 
  pretty_table() |>
  tab_header(title = "Hemifield bias by level by language laterality bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>

## Continuous language (exploratory)
<br>
<br>
`rt_model_ehi <- lmer( rt ~ field*level*ehi*DL_lat*clickhand + (1 | subject) )`
<br>
<br>
```{r}
## Prepare data, so emmeans contrasts show
## Global bias as a positive number.
aah_for_rt_ehi_model <- aah_for_rt_model |>
  mutate(level = level |> factor(levels = c("Local", "Global")),
         field = field |> factor(levels = c("LVF", "RVF")))
```

```{r}
MODEL_DESCRIPTION <- "Model: field * level * ehi * DL_lat * clickhand + (1|subject) + (1|shape)"

if (use_cached_models == FALSE) {
  rt_mm_cont_click <- lmer(
    data = aah_cont,
    formula = rt ~
      field * level * ehi * DL_lat * clickhand
    + (1 | subject)
    + (1 | shape)
  )
  
  saveRDS(rt_mm_cont_click, here(manual_cache_dir, "rt_mm_cont_click.rds"))
} else if (use_cached_models == TRUE) {
  rt_mm_cont_click <- readRDS(here(manual_cache_dir, "rt_mm_cont_click.rds"))
}
```

```{r}
## FxLxH (emmeans)
## Estimate effect of field by level by language lat
rt_emt <- emtrends(rt_mm_cont_click, pairwise ~ field*level, var = "DL_lat",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT
                   )

## Report interaction estimate, with CI and p value.
rt_emt[[1]] |> contrast(interaction = c("consec"), var = "DL_lat") |> 
  summary(infer = T) |>
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = "Field by level by language laterality (cont.) interaction (RT)",
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means local bias is stronger in RVF for right earers (as predicted by LAH), in ms per DL unit (-100 to 100). Multiply this value by 200 to get the estimated difference in LVF global bias for strong left vs. right earers.",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = MODEL_DESCRIPTION,
               locations = cells_title()) |> 
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```

<br>
```{r}
## FxH | L (emmeans)
## Report estimated effect of field, for each level, by ehi.
## Test whether interaction slope differs from zero, with emmeans

## This code estimates the effect of EHI on RT for each field*level:
rt_emt <- emtrends(rt_mm_cont_click, pairwise ~ field*level, var = "DL_lat",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)

## Report interaction estimate, with CI and p value.
rt_emt[[1]] |> contrast("pairwise", by = "level", var = "DL_lat") |> 
  summary(infer = T) |>
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = str_glue("{experiment_label}: Field by language laterality (by Level) (RT)"), 
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means global bias is stronger in LVF for right earers (as predicted by LAH), in ms per DL unit (-100 to 100). Multiply this value by 200 to get the estimated difference in LVF global bias for strong left vs. right earers.",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
br>
