==> analysis/experiments_2-3/analyses/3_rt.Rmd <==
---
title: "Experiment 2: Reaction Time Analyses"
pagetitle: "exp 2 | RT analyses"
author: "Owen Morgan"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: html_document
---

<style type="text/css">
  body{
  font-family: Avenir;
  font-size: 12pt;
}
</style>

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.align = "center", fig.width = 9,
                      fig.height = 6, results = "asis")
options(knitr.kable.NA = "")

cli.progress_show_after <- 0

## Do not use scientific notation until 9 decimal places.
options(scipen = 9, digits = 9)

# Suppress summarise info
options(dplyr.summarise.inform = FALSE)
```

```{r lib}
library(here)
library(tidyverse)
library(cli) # For printing error messages

library(lme4)
library(emmeans)
library(broom)
library(gt)

library(ggpubr) ## functions for plotting SD, SEM, CI.
library(ggbeeswarm)
library(patchwork)
library(ggh4x) ## for nested facets

library(knitr) # For include_graphics


source(here::here("lib", "plots-jepg.R"))
source(here::here("lib", "util.R")) # interaction stats, ...
source(here::here("lib", "demographics.R"))

EXPERIMENT_DIR <- "experiments_2-3"
source(here::here(EXPERIMENT_DIR, "lib", "load_process", "load_process.R"))

## Colors for plots
plot_blue <- c("#337aB7")
plot_colors <- c("#4477AA",
                 "#EE6677",
                 "#CCBB44",
                 "#66CCEE",
                 "#AA3377",
                 "#228833",
                 "#BBBBBB")
```

```{r config}
proc_dir <- set_or_make_dir(label = "proc",
  here::here(EXPERIMENT_DIR, "data", "proc_exp_n1450")
)
fig_dir <- set_or_make_dir(label = "fig",
  here::here(EXPERIMENT_DIR, "figures")
)
manual_cache_dir <- set_or_make_dir(label = "manual cache",
 here::here(EXPERIMENT_DIR, "manual_cache")
)

experiment_label <- "Experiment 2"

## Emmeans settings (use Satterthwaire approx. instead of default t-to-z for large df)
## See: Luke (2017). Evaluating significance in linear mixed-effects models in R
DF_METHOD <- "satterthwaite"
LMERTEST_LIMIT <- 500000

## Figure settings
DEFAULT_HEIGHT = 3
DEFAULT_WIDTH = 5

use_cached_models <- T
use_cached_proc_data <- T
use_cached_figs <- T
```

# {.tabset}
```{r load_data}
## Load "the data" with all subjects & trials.
aah_long <- load_aah_long(proc_dir)

## Load summary data table (for quick demographic analyses)
## with all subjects.
aah_summary_all <- load_aah_summary(proc_dir)
```

```{r prepare_data}
#### PREPARE SUMMARY DATA FOR DEMOGRAPHICS ANALYSES (NAVON EXCLUSIONS)
if (use_cached_proc_data == FALSE) {
cli::cli_alert_info("Filtering summary data for analysis (no excluded subjects).")
cli::cli_bullets(c(" " = "aah_summary_all -> aah_summary"))
aah_summary <- aah_summary_all |>
    filter(exclude_navon == 0) |>
    select(-starts_with("exclude")) |> 
  mutate(
    handedness_extremes = case_when(ehi == -100 ~ "Left",
                                    ehi == 100 ~ "Right",
                                    ehi > -100 & ehi < 100 ~ "Mixed")
  )
  saveRDS(aah_summary, here(manual_cache_dir, "aah_summary.rds"))
} else {
  aah_summary <- readRDS(here(manual_cache_dir, "aah_summary.rds"))
}

#### PREPARE TRIAL-LEVEL DATA FOR RT, ACCURACY ANALYSES
## Prepare "the data" (aah_long) for all analyses:
## Filter out practice trials, absent trials, and excluded subjects
## This data (all present trials, correct or incorrect) will be used for
## accuracy analyses
if (use_cached_proc_data == FALSE) {
cli::cli_alert_info("Filtering long navon data for analysis (no excluded subjects; experimental trials (not practice) only; target present only).")
aah <- aah_long |>
  filter(exclude_navon == 0) |>
  filter(block_type == "main" & target_present == "yes") |>
  select(-starts_with("exclude")) |>
  ## Recode "level", "target" for nicer printing.
  mutate(level = recode(level, global = "Global", local = "Local")) |>
  mutate(target = recode(target, square = "Square", rectangle = "Rectangle")) |>
  ## Code handedness extremes
  mutate(
    handedness_extremes = case_when(ehi == -100 ~ "Left",
                                    ehi == 100 ~ "Right",
                                    ehi > -100 & ehi < 100 ~ "Mixed")
    )
  saveRDS(aah, here(manual_cache_dir, "aah.rds"))
} else {
  aah <- readRDS(here(manual_cache_dir, "aah.rds"))
}

## For RT analyses, prepare dataset with only correct, present trials.
## In our RT model, we only care about correct responses to present trials.
if (use_cached_proc_data == FALSE) {
aah_correct <- aah |> filter(correct == T)
  saveRDS(aah_correct, here(manual_cache_dir, "aah_correct.rds"))
} else {
  aah_correct <- readRDS(here(manual_cache_dir, "aah_correct.rds"))
}

## Relevel field and level with RVF first (unintuitive for plotting),
## so that emmeans will show a positive number for LVF global bias.
if (use_cached_proc_data == FALSE) {
aah_for_rt_model <- aah_correct |>
  mutate(level = level |> factor(levels = c("Global", "Local")),
         field = field |> factor(levels = c("RVF", "LVF")))
  saveRDS(aah_for_rt_model, here(manual_cache_dir, "aah_for_rt_model.rds"))
} else {
  aah_for_rt_model <- readRDS(here(manual_cache_dir, "aah_for_rt_model.rds"))
}

## For accuracy analyses, prepare dataset will all present trials.
## Relevel field and level,
## so that emmeans will show a positive number for LVF global bias.
if (use_cached_proc_data == FALSE) {
aah_for_acc_model <- aah |>
  mutate(level = level |> factor(levels = c("Global", "Local")),
         field = field |> factor(levels = c("RVF", "LVF")))
  saveRDS(aah_for_acc_model, here(manual_cache_dir, "aah_for_acc_model.rds"))
} else {
  aah_for_acc_model <- readRDS(here(manual_cache_dir, "aah_for_acc_model.rds"))
}

#### PREPARE SUBJECT_LEVEL DATA FOR RT, ACCURACY ANALYSES
## Prepare subject-level data (RT)
if (use_cached_proc_data == FALSE) {
rt_subject <- aah_correct |> group_by(subject, field, level, handedness) |>
  summarize(rt = median(rt))
  saveRDS(rt_subject, here(manual_cache_dir, "rt_subject.rds"))
} else {
  rt_subject <- readRDS(here(manual_cache_dir, "rt_subject.rds"))
}


## Prepare subject-level LVF Global bias summary (RT)
if (use_cached_proc_data == FALSE) {
rt_1 <- rt_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = rt) |>
  mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
  mutate(all_one_group = "all_one_group")
  saveRDS(rt_1, here(manual_cache_dir, "rt_1.rds"))
} else {
  rt_1 <- readRDS(here(manual_cache_dir, "rt_1.rds"))
}

## Prepare subject-level LVF Global bias summary (Accuracy)
## (For accuracy, include incorrect trials!)
if (use_cached_proc_data == FALSE) {
acc_subject <- aah |> group_by(subject, field, level, handedness, sample, sample2) |>
  summarize(
    total_responses = n(),
    n_present_resp = sum(correct),
    n_absent_resp = total_responses - n_present_resp,
    n_correct = sum(correct),
    acc = 100 * (n_correct / total_responses)
  ) |>
  select(subject, field, level, acc, handedness, sample, sample2) |>
  mutate(level = recode(level, global = "Global", local = "Local"))
  saveRDS(acc_subject, here(manual_cache_dir, "acc_subject.rds"))
} else {
  acc_subject <- readRDS(here(manual_cache_dir, "acc_subject.rds"))
}

if (use_cached_proc_data == FALSE) {
acc_1 <- acc_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = acc) |>
  mutate(LVF_Global_Bias = (LVF_Global - LVF_Local) - (RVF_Global - RVF_Local)) |>
  mutate(all_one_group = "all_one_group")
  saveRDS(acc_1, here(manual_cache_dir, "acc_1.rds"))
} else {
  acc_1 <- readRDS(here(manual_cache_dir, "acc_1.rds"))
}
```

## Demographics {.tabset .tabset-pills}

Demographics for all included participants.
<br>
```{r demo_summary}
rt_demo <- demo_summary_table(aah_summary)
rt_demo |> pretty_table() |> tab_header(title = "Demographics", subtitle = "Summary")
```
<br>

```{r}
hand_demo_summary_table(aah_summary) |> pretty_table() |>
  tab_footnote("Left: (EHI <= -40)  |  Mixed: (-40 < EHI < 40)  |  Right: (EHI >= 40)")
```
<br>
```{r ehi_distribution_plot_purple1}
middle_purple1 <- "#996F91" ## In between plot colors one and two
fig_path_var <- here(fig_dir, "demo_ehi_dots_purple.png")

if (use_cached_figs == F) {
  ehi_plot_data <- aah_summary |> group_by(ehi) |> summarize(n = n())
  g <- gg_ehi_dotarea(ehi_plot_data, plot_color = middle_purple1,
                      cutoff = 40, cutoff_label_ypos = 120)
  ggsave(fig_path_var, g, "png", height = 3, width = 6)
}

include_graphics(fig_path_var)
```


## Categorical handedness (primary) {.tabset .tabset-pills}

### Summary

*When handedness is binned as left (EHI <= -40) or right (EHI > +40):*
<br>
*Do we find an interaction of field x level x handedness?*
<br>
*At the local level, do left handers show significantly different hemifield bias than right handers? (Local: field x handedness)*
<br>
*At the local level, what is each group's field bias? (Local: By handedness group, effect of field)*
<br>

For reaction time, LVF global>local bias was significantly reduced in left handers (\textit{n} = 529, EHI <= -40) compared to right handers (\textit{n }= 413, EHI >= +40; difference between groups = 20.51ms, 95\% CI [10.45, 30.57], $\chi^2(1)$ = 15.98, \textit{p} < .0001, two-sided). Both groups showed significant LVF global>local bias: for right handers, the effect size was 44.09ms (95\% CI [36.54, 51.63], \textit{$t$}(111,144.6) = 11.46, \textit{p} = < .0001, two-sided); for left handers, 23.57ms (95\% CI [16.92, 30.23], \textit{$t$}(111,144.1) = 6.94, \textit{p} < .0001, two-sided.

within the local level, left handers showed reduced hemifield bias, in the direction predicted by AAH: right handers responded faster to local stimuli in the RVF than LVF by 18.16ms (95\% CI [12.75, 23.57], \textit{$t$}(111,145.9) = 6.58, \textit{p} <.0001, two-sided), whereas left handers responded (insignificantly) faster to local stimuli in the LVF than RVF by 0.56ms (95\% CI [-4.21, 5.33], \textit{$t$}(111,145.6) = 0.23, \textit{p} = .82, two-sided; difference = 18.72ms, 95\% CI [11.51, 26.83], \textit{$t$}(111,145.78) = 5.09, \textit{p} <.0001, two-sided).

### Stats
Reaction time is modeled as a linear effect of field, level, and handedness, using data from every target-present trial with a "go" response:
<br>
<br>
`lmer( rt ~ field*level*handedness + (1 | subject) )`
<br>
<br>
```{r rt_model_2bins}
#### FxLxH (prepare model)
## Make a model with two handedness bins: Right and Left
if (use_cached_models == FALSE) {
  aah_for_rt_model_2bins <- aah_for_rt_model |> filter(handedness %in% c("Right", "Left"))
  rt_model_2bins <- lmer(
    rt ~ field:level:handedness
    + field:level
    + field:handedness
    + level:handedness
    + field + level + handedness
    + (1 | subject),
    data = aah_for_rt_model_2bins
    )
  
  ## Create emmeans model object, and manually cache it.
  rt_emm_2bins <- emmeans(rt_model_2bins, ~ field*level*handedness,
                          lmer.df = DF_METHOD,
                          lmerTest.limit = LMERTEST_LIMIT)
  rt_interaction_emm <- rt_emm_2bins |> 
  contrast(interaction = c("consec")) |>
  summary(infer = T)
  
  ## Manually cache model
  saveRDS(rt_model_2bins, here(manual_cache_dir, "rt_model_2bins.rds"))
  saveRDS(rt_emm_2bins, here(manual_cache_dir, "rt_emm_2bins.rds"))
  
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_model_2bins <- readRDS(here(manual_cache_dir, "rt_model_2bins.rds"))
  rt_emm_2bins <- readRDS(here(manual_cache_dir, "rt_emm_2bins.rds"))
}
```
<br>
```{r}
### FxLxH (Likelihood ratio test)
if (use_cached_models == FALSE) {
## Use anova() on competing models to test 3-way interaction.
rt_model_no_interaction <- update(rt_model_2bins, . ~ . - field:level:handedness)
interaction_anova <- interaction_stats(rt_model_2bins, rt_model_no_interaction)
  saveRDS(interaction_anova, here(manual_cache_dir, "rt_mm_cat_anova.rds"))
} else if (use_cached_models == TRUE) {
  interaction_anova <- readRDS(here(manual_cache_dir, "rt_mm_cat_anova.rds"))
}

interaction_anova |>
  as_tibble() |>
  rename(p.value = `Pr(>Chisq)`) |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = str_glue("{experiment_label}: Field by Level by Handedness (RT)"), 
             subtitle = "ANOVA: compare models with vs. without interaction term") |> 
  tab_footnote(footnote = "F-test (two-sided? https://daniellakens.blogspot.com/2016/04/one-sided-f-tests-and-halving-p-values.html)",
               locations = cells_column_labels(columns = p.value))
```

<br>
```{r}
### FxLxH
## Use emmeans() to test 3-way interaction.
if (use_cached_models == FALSE) {
rt_interaction_emm <- rt_emm_2bins |> 
  contrast(interaction = c("consec")) |>
  summary(infer = T)
  saveRDS(rt_interaction_emm, here(manual_cache_dir, "rt_emm_2bins_FLH.rds"))
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_interaction_emm <- readRDS(here(manual_cache_dir, "rt_emm_2bins_FLH.rds"))
}

rt_interaction_emm |>
  as_tibble() |>
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = str_glue("{experiment_label}: Field by Level by Handedness (RT)"),
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means LVF global bias is stronger in right handers (as predicted by AAH)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```

<br>
```{r}
#### FxL | H
## Estimate the effect of field by level for each handedness group
if (use_cached_models == FALSE) {
rt_emm2 <- emmeans(rt_model_2bins, ~field*level | handedness, lmer.df = DF_METHOD,
                          lmerTest.limit = LMERTEST_LIMIT) |>
  contrast(interaction = "consec") |>
  summary(infer = T, adj = "none")
  saveRDS(rt_emm2, here(manual_cache_dir, "rt_emm_2bins_FL_H.rds"))
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_emm2 <- readRDS(here(manual_cache_dir, "rt_emm_2bins_FL_H.rds"))
}

rt_emm2 |>
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = str_glue("{experiment_label}: Field by Level (by Handedness) (RT)")) |> 
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
#### FxH | L
if (use_cached_models == FALSE) {
## Estimate the effect of field for each level for each handedness group
rt_emm_FH_L <- rt_emm_2bins |> contrast(interaction = "consec", by = c("level")) |> 
  summary(infer = T) 
  saveRDS(rt_emm_FH_L, here(manual_cache_dir, "rt_emm_2bins_FL_H.rds"))
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_emm_FH_L <- readRDS(here(manual_cache_dir, "rt_emm_2bins_FL_H.rds"))
}

rt_emm_FH_L |> 
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = str_glue("{experiment_label}: Field by Handedess (by Level) (RT)")) |> 
  tab_footnote(footnote = "A positive number means more RVF bias for right handers",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))

```


```{r}
#### F | L | H
if (use_cached_models == FALSE) {
## Estimate the effect of field for each level for each handedness group
rt_emm2 <- emmeans(rt_model_2bins, ~field | handedness + level, lmer.df = DF_METHOD,
                          lmerTest.limit = LMERTEST_LIMIT) |>
  contrast("revpairwise", by = c("handedness","level")) |>
  summary(infer = T, adj = "none")
  saveRDS(rt_emm2, here(manual_cache_dir, "rt_emm_2bins_F_L_H.rds"))
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_emm2 <- readRDS(here(manual_cache_dir, "rt_emm_2bins_F_L_H.rds"))
}

rt_emm2 |>
  as_tibble() |> 
  format_p.value() |> 
  arrange(handedness, level) |> 
  pretty_table() |>
  tab_header(title = "LVF Global bias by handedness bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```

### Plots

Diamonds and lineranges show mixed-effects model point estimates and 95% CI.

```{r plot_cat_model}
## Start over, with numeric y axis
## Add a line with CI from the model
fig_path_var <- here(fig_dir, "rt_1_lmer.png")

if (use_cached_figs == F) {
rt_emm <- readRDS(here(manual_cache_dir, "rt_emm_2bins.rds"))

## Get estimates for right and left handers.
rt_emm |>
  contrast("pairwise", by = c("handedness", "level")) |>
  summary(infer = T, level = .95)

emm_data_2 <- rt_emm |>
  contrast("pairwise", by = c("handedness", "level")) |>
  summary(infer = T, level = .95) |>
  as_tibble() |> 
  ## Fix units
  mutate(
    estimate = estimate * -1,
    lower.CL = lower.CL * -1,
    upper.CL = upper.CL * -1
  )


plot_data <- emm_data_2 |> mutate(dv = estimate)

## Extract n's (MANUALLY COUNT N IN DATAFRAME USED IN MODEL)
n_left <- rt_1 |> group_by(handedness) |> count() |> filter(handedness == "Left") |> pull(n)
n_right <- rt_1 |> group_by(handedness) |> count() |> filter(handedness == "Right") |> pull(n)

g <- gg_rt_2_lmer(
  title = "Hemifield bias by level (EHI cut at +/-40)",
  plot_data,
  handedness_labeller = NULL,
  plot_colors = plot_colors,
  direction_labels = list(right = "RVF bias",
                          left = "LVF bias"),
  direction_labels_pos = list(right = 38,
                              left = -38),
  xlims = list(upper = 38,
               lower = -38),
  xbreaks = list(major = 10,
                 minor = 5),
  n_subjects = list(
    right = n_right,
    left = n_left
  )
)

  ggsave(fig_path_var, g, "png", height = DEFAULT_HEIGHT, width = DEFAULT_WIDTH)
}

include_graphics(fig_path_var)
```
<br>


## Categorical handedness (exploratory) {.tabset .tabset-pills}

### Summary

*When handedness is binned as left (EHI <= -100) or right (EHI > +100):*
<br>
*Do we find an interaction of field x level x handedness?*
<br>
*At the local level, do left handers show significantly different hemifield bias than right handers? (Local: field x handedness)*
<br>
*At the local level, what is each group's field bias? (Local: By handedness group, effect of field)*
<br>

Limiting analysis to strong left and right handers with EHI scores of +/-100, left handers' LVF>RVF global bias was reduced by 17.44ms (95\% CI [1.50, 33.38], $\chi^2(1)$ = 4.6, \textit{p} = .032, two-sided). Both groups showed significant LVF global>local bias: for right handers, the effect size was 36.96ms (95\% CI [25.36 ,48.56], \textit{$t$}(44,479.7) = 6.25, \textit{p} < .0001, two-sided); for left handers, 19.52ms (95\% CI [8.58, 30.45], \textit{$t$}(44,479.0) = 3.50, \textit{p} = .0005, two-sided).

within the local level, right and left handers showed reversed hemifield bias: right handers responded faster to local stimuli in the RVF than LVF by 9.97ms (95\% CI [1.61, 18.32],  \textit{$t$}(44,479.9) = 2.34, \textit{p} = .019, two-sided), whereas left handers responded faster to local stimuli in the LVF than RVF by 7.98ms (95\% CI [0.14, 15.83], \textit{$t$}(44,479.4) = 2.00 \textit{p} = .046, two-sided; difference = 17.95ms, 95\% CI [6.48, 29.41], \textit{$t$}(44,479.7) = 3.07, \textit{p} = .002, two-sided).


### Stats

Reaction time is modeled as a linear effect of field, level, and handedness, using data from every target-present trial with a "go" response:
<br>
<br>
`lmer( rt ~ field*level*handedness + (1 | subject) )`
<br>
<br>
```{r rt_model_zoom}
#### FxLxH (prepare model)
## Make a model with two handedness bins: Right and Left
if (use_cached_models == FALSE) {
  aah_for_rt_model_zoom <- aah_for_rt_model |> filter(handedness_extremes %in% c("Right", "Left"))
  rt_model_zoom <- lmer(
    rt ~ field:level:handedness
    + field:level
    + field:handedness
    + level:handedness
    + field + level + handedness
    + (1 | subject),
    data = aah_for_rt_model_zoom
    )
  
  ## Create emmeans model object, and manually cache it.
  rt_emm_zoom <- emmeans(rt_model_zoom, ~ field*level*handedness,
                          lmer.df = DF_METHOD,
                          lmerTest.limit = LMERTEST_LIMIT)
  rt_interaction_emm <- rt_emm_zoom |> 
  contrast(interaction = c("consec")) |>
  summary(infer = T)
  
  ## Manually cache model
  saveRDS(rt_model_zoom, here(manual_cache_dir, "rt_model_zoom.rds"))
  saveRDS(rt_emm_zoom, here(manual_cache_dir, "rt_emm_zoom.rds"))
  
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_model_zoom <- readRDS(here(manual_cache_dir, "rt_model_zoom.rds"))
  rt_emm_zoom <- readRDS(here(manual_cache_dir, "rt_emm_zoom.rds"))
}
```
<br>
```{r}
### FxLxH (Likelihood ratio test)
if (use_cached_models == FALSE) {
## Use anova() on competing models to test 3-way interaction.
rt_model_no_interaction <- update(rt_model_zoom, . ~ . - field:level:handedness)
interaction_anova <- interaction_stats(rt_model_zoom, rt_model_no_interaction)
  saveRDS(interaction_anova, here(manual_cache_dir, "rt_mm_cat_anova_zoom.rds"))
} else if (use_cached_models == TRUE) {
  interaction_anova <- readRDS(here(manual_cache_dir, "rt_mm_cat_anova_zoom.rds"))
}

interaction_anova |>
  as_tibble() |>
  rename(p.value = `Pr(>Chisq)`) |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = str_glue("{experiment_label}: Field by Level by Handedness (RT)"), 
             subtitle = "ANOVA: compare models with vs. without interaction term") |> 
  tab_footnote(footnote = "F-test (two-sided? https://daniellakens.blogspot.com/2016/04/one-sided-f-tests-and-halving-p-values.html)",
               locations = cells_column_labels(columns = p.value))
```
<br>
```{r rt_interaction_emm}
### FxLxH
if (use_cached_models == FALSE) {
rt_interaction_emm <- rt_emm_zoom |> 
  contrast(interaction = c("consec")) |>
  summary(infer = T)
  saveRDS(rt_interaction_emm, here(manual_cache_dir, "rt_emm_FLH_zoom.rds"))
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_interaction_emm <- readRDS(here(manual_cache_dir, "rt_emm_FLH_zoom.rds"))
}

rt_interaction_emm |>
  as_tibble() |>
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = str_glue("{experiment_label}: Field by Level by Handedness (RT)"),
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means LVF global bias is stronger in right handers (as predicted by AAH)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```

<br>
```{r rt_handedness_bias}
#### FxL | H
##TODO: fix report show that the correct table is shown here.
## Estimate the effect of field by level for each handedness group
if (use_cached_models == FALSE) {
rt_emm2 <- emmeans(rt_model_zoom, ~field*level | handedness, lmer.df = DF_METHOD,
                          lmerTest.limit = LMERTEST_LIMIT) |>
  contrast(interaction = "consec") |>
  summary(infer = T, adj = "none")
  saveRDS(rt_emm2, here(manual_cache_dir, "rt_emm_zoom_FL_H.rds"))
  
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_emm2 <- readRDS(here(manual_cache_dir, "rt_emm_zoom_FL_H.rds"))
}

rt_emm2 |>
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |>
  tab_header(title = str_glue("{experiment_label}: Field by Level (by Handedness) (RT)")) |> 
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
#### FxH | L
if (use_cached_models == FALSE) {
rt_emm_FH_L <- rt_emm_zoom |> contrast(interaction = "consec", by = c("level")) |> 
  summary(infer = T)
  
  saveRDS(rt_emm_FH_L, here(manual_cache_dir, "rt_emm_FH_L_zoom.rds"))
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_emm_FH_L <- readRDS(here(manual_cache_dir, "rt_emm_FH_L_zoom.rds"))
}

rt_emm_FH_L |> 
  as_tibble() |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = str_glue("{experiment_label}: Field by Handedess (by Level) (RT)")) |> 
  tab_footnote(footnote = "A positive number means more RVF bias for right handers",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))

```


```{r rt_handedness_bias_bylevel}
#### F | L | H
if (use_cached_models == FALSE) {
## Estimate the effect of field for each level for each handedness group
rt_emm2 <- emmeans(rt_model_zoom, ~field | handedness + level, lmer.df = DF_METHOD,
                          lmerTest.limit = LMERTEST_LIMIT) |>
  contrast("revpairwise", by = c("handedness","level")) |>
  summary(infer = T, adj = "none")
  saveRDS(rt_emm2, here(manual_cache_dir, "rt_emm_zoom_F_L_H.rds"))
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_emm2 <- readRDS(here(manual_cache_dir, "rt_emm_zoom_F_L_H.rds"))
}

rt_emm2 |>
  as_tibble() |> 
  format_p.value() |> 
  arrange(handedness, level) |> 
  pretty_table() |>
  tab_header(title = "LVF Global bias by handedness bin (RT)") |>
  tab_footnote(footnote = "A positive number means global bias (faster RT for global)",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Confidence level: 95%",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Two-sided, uncorrected",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```



### Plots

```{r plot_cat_model_zoom}
## Categorical plots: EHI extremes
## Start over, with numeric y axis
## Add a line with CI from the model
fig_path_var <- here(fig_dir, "rt_1_lmer_zoom.png")

if (use_cached_figs == F) {
## Get estimates for right and left handers.
zoom_data <- aah_for_rt_model |>
  filter(handedness_extremes %in% c("Right", "Left"))

emm_data_2 <- rt_emm_zoom |>
  contrast("pairwise", by = c("handedness", "level")) |>
  summary(infer = T, level = .95) |>
  as_tibble() |>
  ## Fix units
  mutate(
    estimate = estimate * -1,
    lower.CL = lower.CL * -1,
    upper.CL = upper.CL * -1
  )


plot_data <- emm_data_2 |> mutate(dv = estimate)

## Extract n's (MANUALLY COUNT N IN DATAFRAME USED IN MODEL)
n_tbl <- aah_for_rt_model |> group_by(subject) |>
  summarize(subject = first(subject), handedness_extremes = first(handedness_extremes)) |> group_by(handedness_extremes) |> count()
n_left <- n_tbl |> filter(handedness_extremes == "Left") |> pull(n)
n_right <- n_tbl |> filter(handedness_extremes == "Right") |> pull(n)

g <- gg_rt_2_lmer(
  title = "Hemifield bias by level (EHI +/-100)",
  plot_data,
  handedness_labeller = NULL,
  plot_colors = plot_colors,
  direction_labels = list(right = "RVF bias",
                          left = "LVF bias"),
  direction_labels_pos = list(right = 38,
                              left = -38),
  xlims = list(upper = 38,
               lower = -38),
  xbreaks = list(major = 10,
                 minor = 5),
  n_subjects = list(
    right = n_right,
    left = n_left
  )
)

  ggsave(fig_path_var, g, "png", height = DEFAULT_HEIGHT, width = DEFAULT_WIDTH)
}

include_graphics(fig_path_var)
```


## Continuous handedness (primary) {.tabset .tabset-pills}

### Summary
Do we find an interaction of field x level x handedness (continuous EHI score)?
<br>
<br>

With handedness treated as continuous, we found that degree of left handedness predicted reduced LVF global>local bias (0.13ms per EHI unit, 95\% CI [0.06, 0.19], $\chi^2(1)$ = 17.88, \textit{p} < .0001, two-sided). Estimated LVF global>local bias for EHI +100 (strong right handers) was 45.61ms (95\% CI [37.73, 53.48]), and for EHI -100 (strong left handers), 19.92ms (95\% CI [12.72, 27.12]), a difference of 25.69ms.

within the local level, left handedness predicted reduced RVF bias (0.12ms per EHI unit, 95\% CI [0.07, 0.18], \textit{$t$}(131,342.5) = 5.65, \textit{p} = < .0001, two-sided). Strong right handers (EHI = +100) showed estimated RVF bias of 21.36ms (95\% CI [15.70, 27.01]), and strong left handers (EHI = -100) showed estimated LVF bias of 3.28ms (95\% CI [-1.88, 8.43]), a difference of 24.54ms.

### Stats

```{r}
rt_ehi_1 <- rt_1 |>
  left_join(aah_summary |> select(subject, ehi)) |>
  select(subject, ehi, LVF_Global_Bias)
```

Model RT as a linear effect of field, level, and EHI (continuous):
<br>
<br>
`rt_model_ehi <- lmer( rt ~ field*level*ehi + (1 | subject) )`
<br>
<br>
```{r rt_model_ehil}
## Prepare data, so emmeans contrasts show
## Global bias as a positive number.
aah_for_rt_ehi_model <- aah_for_rt_model |>
  mutate(level = level |> factor(levels = c("Local", "Global")),
         field = field |> factor(levels = c("LVF", "RVF")))
```


```{r rt_model_ehi_caching}
if (use_cached_models == FALSE) {
  rt_model_ehi <- lmer(rt ~ field:level:ehi + field:level + field:ehi + level:ehi + field + level + ehi + (1 | subject), data = aah_for_rt_ehi_model)
  
  ## Create emmeans model object, and manually cache it.
  rt_emm_ehi <- emmeans(rt_model_ehi, ~ field * level * ehi,
                        lmer.df = DF_METHOD,
                        lmerTest.limit = LMERTEST_LIMIT)
  
  ## Manually cache model
  saveRDS(rt_model_ehi, here(manual_cache_dir, "rt_model_ehi.rds"))
  saveRDS(rt_emm_ehi, here(manual_cache_dir, "rt_emm_ehi.rds"))
  
} else if (use_cached_models == TRUE) {
  ## Load cached model
  rt_model_ehi <- readRDS(here(manual_cache_dir, "rt_model_ehi.rds"))
  rt_emm_ehi <- readRDS(here(manual_cache_dir, "rt_emm_ehi.rds"))
}
```

```{r rt_ehi_interaction_anova, echo = F}
#### FxLxH (anova)
## Use anova() on competing models to test 3-way interaction.
if (use_cached_models == FALSE) {
rt_model_no_interaction <- update(rt_model_ehi, . ~ . - field:level:ehi)
interaction_anova <- interaction_stats(rt_model_ehi, rt_model_no_interaction)

  saveRDS(interaction_anova, here(manual_cache_dir, "rt_mm_cont_anova.rds"))
} else if (use_cached_models == TRUE) {
  interaction_anova <- readRDS(here(manual_cache_dir, "rt_mm_cont_anova.rds"))
}
interaction_anova |>
  as_tibble() |>
  rename(p.value = `Pr(>Chisq)`) |> 
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = str_glue("{experiment_label}: Field by Level by EHI (RT)"), 
             subtitle = "ANOVA: compare models with vs. without interaction term") |> 
  tab_footnote(footnote = "F-test (two-sided? https://daniellakens.blogspot.com/2016/04/one-sided-f-tests-and-halving-p-values.html)",
               locations = cells_column_labels(columns = p.value))
```
<br>
```{r}
#### FxLxH (emmeans)
## Test whether interaction slope differs from zero, with emmeans

## This code estimates the effect of EHI on RT for each field*level:
rt_emt <- emtrends(rt_model_ehi, pairwise ~ field*level, var = "ehi", lmer.df = DF_METHOD,
                        lmerTest.limit = LMERTEST_LIMIT)

## Report interaction estimate, with CI and p value.
rt_emt[[1]] |> contrast(interaction = c("consec"), var = "ehi") |> 
  summary(infer = T) |>
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = str_glue("{experiment_label}: Field by Level by EHI (RT)"), 
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means global bias is stronger in LVF for right handers (as predicted by AAH), in ms per EHI unit (-100 to 100). Multiply this value by 200 to get the estimated difference in LVF global bias for strong left vs. right handers.",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
<br>
```{r}
## Report estimated effect of field, for each level, by ehi.
## Test whether interaction slope differs from zero, with emmeans

## This code estimates the effect of EHI on RT for each field*level:
rt_emt <- emtrends(rt_model_ehi, pairwise ~ field*level, var = "ehi",
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)

## Report interaction estimate, with CI and p value.
rt_emt[[1]] |> contrast("pairwise", by = "level", var = "ehi") |> 
  summary(infer = T) |>
  format_p.value() |> 
  pretty_table() |> 
  tab_header(title = str_glue("{experiment_label}: Field by EHI (by Level) (RT)"), 
             subtitle = "Compare effect estimate to zero with emmeans()") |>
  tab_footnote(footnote = "A positive number means global bias is stronger in LVF for right handers (as predicted by AAH), in ms per EHI unit (-100 to 100). Multiply this value by 200 to get the estimated difference in LVF global bias for strong left vs. right handers.",
               locations = cells_column_labels(columns = estimate)) |>
  tab_footnote(footnote = "Two-sided",
               locations = cells_column_labels(columns = p.value)) |>
  tab_footnote(footnote = "Confidence level: 95",
               locations = cells_column_labels(columns = ends_with("CL"))) |>
  tab_footnote(footnote = "Z-approximation",
               locations = cells_column_labels(columns = df))
```
br>

```{r}
## LVF Global bias by EHI.
## Pull estimates for strong left and right handers (ehi = -100, +100)
## Estimate effect of ehi on field x level.

## Use ref_grid (following quant 2 HW5)
ref <- ref_grid(rt_model_ehi, at = list(ehi = c(-100, 0, 100)),
                   lmer.df = DF_METHOD, lmerTest.limit = LMERTEST_LIMIT)

## Show estimates for left and right handers.
ref |>
  contrast(interaction = c("consec"), by = "ehi") |>
  summary(infer = T, type = "response") |>
  as_tibble() |>
  format_p.value() |>
  pretty_table() |>
  tab_header(title = "Estimated Field by Level interaction, by EHI score") |>
  tab_footnote(footnote = "Strong left hander: -100; Mixed hander: 0; Strong right hander: +100",
               locations = cells_column_labels(columns = ehi)) |>
  tab_footnote(footnote = "Estimated LVF global bias (ms); a positive number means LVF global bias",
               locations = cells_column_labels(columns = estimate))
```
<br>
```{r}
## Global bias by field, by handedness.
ref |>
  contrast("pairwise", by = c("level", "ehi")) |>
  summary(infer = T, type = "response") |>
  as_tibble() |>
  # filter(contrast %in% c("LVF Local - RVF Local", "LVF Global - RVF Global")) |>
  format_p.value() |>
  arrange(ehi) |>
  pretty_table() |>
  tab_header(title = "Estimated global bias by field, by EHI score") |>
  tab_footnote(footnote = "Estimated global bias (ms); a positive number means global bias",
               locations = cells_column_labels(columns = estimate))
```

### Plots
```{r plot_rt_cor_bins_config}
# plot_color = plot_blue
middle_purple1 <- "#996F91" ## In between plot colors one and two
middle_purple2 <- "#886e92" ## In between plot colors one and two
plot_color <- middle_purple1

h_plot_colors <- c(plot_colors[[1]], middle_purple1, plot_colors[[2]])
```

```{r}
## Plot correlation for local and global shapes (facets)
## Add a line with CI from the model
fig_path_var <- here(fig_dir, "rt_2_cor_bins_line.png")
fig_path_var_wide <- here(fig_dir, "rt_2_cor_bins_line.png")


if (use_cached_figs == F) {
## Extract line of best fit, confidence bounds from model.
rt_model_ehi <- readRDS(here(manual_cache_dir, "rt_model_ehi.rds"))


line_data <- rt_model_ehi |> 
  ref_grid(at = list(ehi = seq(-100, 100, 12.5)),
           lmer.df = DF_METHOD,
                        lmerTest.limit = LMERTEST_LIMIT
           ) |> 
  emmeans(~ field * ehi, by = "level") |> 
  contrast(interaction = c("consec"), by = c("ehi", "level"), level = .95) |> 
  summary(infer = T) |> 
  as_tibble()

rt_subject_plot <- rt_subject |>
  pivot_wider(names_from = c(field, level),
              values_from = rt) |>
  mutate(LVF_Global_Bias = (RVF_Global - RVF_Local) - (LVF_Global - LVF_Local)) |>
  left_join(aah_summary, by = c("subject", "handedness")) |>
  select(subject, ehi, starts_with("LVF"), handedness, handedness_extremes) |>
  ## Set "dv" here to make it easy to write a reusable function
  mutate(dv = LVF_Global_Bias) |>
  ungroup()

rt_subject_plot <- rt_subject |>
  pivot_wider(names_from = c(field),
              values_from = rt) |>
   mutate(LVF_Bias = RVF - LVF) |> 
  left_join(aah_summary, by = c("subject", "handedness")) |>
  select(subject, level, ehi, starts_with("LVF"), handedness, handedness_extremes) |> 
  mutate(dv = LVF_Bias) |> 
  ungroup()

  g <-
    gg_rt_2_line(
      title = "Frequency specialization by handedness (continuous)",
      y_title = "RVF - LVF reaction time (ms)",
      rt_subject_plot = rt_subject_plot,
      plot_colors = h_plot_colors,
      plot_color = plot_color,
      group_by_level = T,
      direction_labels = list(
        up = "LVF Bias",
        down = "RVF Bias"),
      direction_labels_pos = list(
        up = 48,
        down = -68),
      ylims = list(
        upper = 50,
        lower = -70),
      ybreaks = list(
        major = 10,
        minor = 5)
    ) |>
    gg_style_cor_bin()
  
  g <- g +
  geom_line(data = line_data,
            aes(y = estimate)) +
    geom_ribbon(data = line_data,
                inherit.aes = F,
              aes(x = ehi, ymin = lower.CL, ymax = upper.CL),
              fill = plot_color,
              alpha = .2)
  
  g <- g + facet_wrap(~ level, nrow = 1)
  ggsave(fig_path_var, g, "png", height = 4, width = 4)
  ggsave(fig_path_var_wide, g, "png", height = 4, width = 8)
}

include_graphics(fig_path_var)
```
<br>
